refactor: 컴포넌트 스플릿팅, cartUtils 함수 재사용
